nvidia opencl fixes

+++ include/dehancer/gpu/kernels/opencl/common.h
@@ -9,7 +9,6 @@ static __constant float3 kIMP_Y_YUV_factor = {0.2125, 0.7154, 0.0721};
 
 __constant sampler_t linear_normalized_sampler = CLK_NORMALIZED_COORDS_TRUE | CLK_ADDRESS_CLAMP_TO_EDGE | CLK_FILTER_LINEAR;
 __constant sampler_t nearest_sampler = CLK_NORMALIZED_COORDS_FALSE | CLK_ADDRESS_CLAMP_TO_EDGE | CLK_FILTER_NEAREST;
-//__constant sampler_t nearest__sampler = CLK_NORMALIZED_COORDS_FALSE | CLK_ADDRESS_CLAMP | CLK_FILTER_NEAREST;
 
 static inline float4 sampledColor(
         __read_only image2d_t inTexture,
@@ -53,8 +52,8 @@ __kernel void grid_kernel(int levels, __write_only image2d_t destination )
 
   int2 gid = (int2)(x, y);
 
-  float2 coords = (float2)((float)gid.x / (w - 1),
-                           (float)gid.y / (h - 1));
+  float2 coords = (float2)((float)gid.x / (float)(w - 1),
+                           (float)gid.y / (float)(h - 1));
 
   int num = levels*2;
   int index_x = (int)(coords.x*(num));
